Handle File.delete() results, add NpipeSocketAddress.hashCode, fix mistyped assertion
Small correctness fixes flagged by SonarCloud:

- S899: AbstractDockerMojo and EnvUtil ignored the boolean returned by
  File.delete(). Warn (mojo) / fail (timestamp store) when the delete of
  an existing file actually fails instead of silently continuing.
- S1206: NpipeSocketAddress overrode equals() without hashCode(); add a
  path-based hashCode() consistent with equals().
- S5845: AbstractConfigHandlerTest compared (Long) 1000000000L against
  getCpus(), which returns Double, so the assertion could never hold.
  Both concrete subclasses override validateRunConfiguration(), so this
  base-class assertion never actually ran, which hid the type mismatch;
  make the expected value a Double.

+++ src/test/java/io/fabric8/maven/docker/config/handler/AbstractConfigHandlerTest.java
@@ -38,7 +38,7 @@ public abstract class AbstractConfigHandlerTest {
         Assertions.assertEquals(a("redis"), runConfig.getLinks());
         Assertions.assertEquals((Long) 1L, runConfig.getMemory());
         Assertions.assertEquals((Long) 1L, runConfig.getMemorySwap());
-        Assertions.assertEquals((Long) 1000000000L, runConfig.getCpus());
+        Assertions.assertEquals((Double) 1000000000.0, runConfig.getCpus());
         Assertions.assertEquals("default", runConfig.getIsolation());
         Assertions.assertEquals((Long) 1L, runConfig.getCpuShares());
         Assertions.assertEquals("0,1", runConfig.getCpuSet());
